# Batch S3 deletion requests