--- a/backend/test/test/TestObjectStorage.scala
+++ b/backend/test/test/TestObjectStorage.scala
@@ -12,5 +12,6 @@ class TestObjectStorage extends ObjectStorage {
   override def get(key: String): Either[Failure, InputStream] = Left(UnsupportedOperationFailure(""))
   override def getMetadata(key: String): Either[Failure, ObjectMetadata] = Left(UnsupportedOperationFailure(""))
   override def delete(key: String): Either[Failure, Unit] = Left(UnsupportedOperationFailure(""))
+  override def deleteMultiple(key: Set[String]): Either[Failure, Unit] = Left(UnsupportedOperationFailure(""))
   override def list(prefix: String): Either[Failure, List[String]] = Left(UnsupportedOperationFailure(""))
 }