After an excellent pairing session with @rtyley, we realised that we could get some performance gains by sending multiple keys to delete in a single S3 call, rather than making multiple calls.

This means we make at most two delete requests to S3: one to delete everything from the preview bucket (page previews and the whole-document preview), and one the delete the blob from the ingest bucket.

The biggest gains would be in documents with many pages since before we were making at least one call per page, e.g. 324 requests become 2 requests.

But even for a document with one page, we still reduce the S3 requests from 3 to 2.

# local perf stats
## before (for 322 page "Functional Programming in Scala" book)
**900ms**
```
Delete page previews: took 0 minutes 0 seconds 780 milliseconds
Preview S3 delete: took 0 minutes 0 seconds 5 milliseconds
Ingest S3 delete: took 0 minutes 0 seconds 5 milliseconds
Delete blob from neo4j: took 0 minutes 0 seconds 59 milliseconds
Delete blob from elasticsearch: took 0 minutes 0 seconds 18 milliseconds
Total to delete resource: took 0 minutes 0 seconds 900 milliseconds
```

## after
**145ms**
```
Get page preview S3 keys: took 0 minutes 0 seconds 54 milliseconds
Deleting 323 objects from preview storage
Preview storage S3 delete: took 0 minutes 0 seconds 61 milliseconds
Ingest storage S3 delete: took 0 minutes 0 seconds 4 milliseconds
Delete blob from neo4j: took 0 minutes 0 seconds 12 milliseconds
Delete blob from elasticsearch: took 0 minutes 0 seconds 5 milliseconds
Total to delete resource: took 0 minutes 0 seconds 145 milliseconds
```

# prod perf stats
## before
Note: these may be distorted by certain subsets of the documents in the collection I'm deleting having different properties. I've seen periods in which deletions went much quicker, presumably because it had a run of documents that had no page previews. I don't understand the order in which documents come back from the elasticsearch so don't know how much this clustering will tend to occur.

Averaged [500 deletions](http://localhost:5601/app/discover#/?_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2022-10-25T07:36:58.871Z',to:'2022-10-25T09:36:27.579Z'))&_a=(columns:!(message),filters:!(),index:d2720940-0e30-11eb-a809-5706b391daea,interval:auto,query:(language:kuery,query:'%22Total%20to%20delete%20resource%22'),sort:!())): 263 ms total